This commit is all about wall detection.
Fortunately, Roii understands the Cartesian coordinates system so he
quickly understood why we need to keep track of the tank's (x,y)
coordinates.  As a preliminary step, we converted the series of
Console.WriteLine calls to a single map string.
Whenever the tank makes a move, we refer back to this map, using the
tank's coordinates and check if the "future" move places the tank on the
road (represented by spaces), or on a wall (represented by the pound
sign).
Converting the 2-dimensional (x,y) coordinates to a 1-dimensional
coordinate on the 'map' string took several explanations.  He seemed to
understand the concept, but it is far from trivial.  I also explained a
bit about the memory representation of strings.
As usual, I worked with him, showing him how to write the code for
wall-detection when moving right and he completed the code for the other
three directions.
We also added a display of the tank's (x,y) coordinates at the bottom of
the screen, which he thought was cool.  As before, we worked out the
solution while discussing possible solutions, experimenting, fixing and
reiterating.